auth-utils/pyproject.toml
[project]
name = "auth-utils"
version = "0.2.0"
description = "Shared authentication and authorization utilities for MCP Gateway Registry"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "pydantic>=2.11.3",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "pyjwt>=2.6.0",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
auth_utils = ["py.typed"]

# Tasks of this specific workspace member should be run from the member directory, i.e. `./auth-utils`.
[tool.poe.tasks]
# Run tests
test = { shell = "uv run pytest tests -v" }
test-cov = { shell = "uv run pytest tests -v --cov=auth_utils --cov-report=html --cov-report=term --cov-report=xml" }
